Moved Question: The code works
