+++ profiles_api/admin.py
@@ -4,11 +4,12 @@ from profiles_api import models
 
 from profiles_api.topic.topic_model import Topic
 from profiles_api.subtopic.subtopic_model import Subtopic
+from profiles_api.question.question_model import Question
 
 
 admin.site.register(models.UserProfile)
 admin.site.register(models.ProfileFeedItem)
-admin.site.register(models.Question)
+admin.site.register(Question)
 admin.site.register(Topic)
 admin.site.register(Subtopic)
 admin.site.register(models.Answer)
