+++ profiles_api/serializers.py
@@ -41,17 +41,6 @@ class ProfileFeedItemSerializer(serializers.ModelSerializer):
         extra_kwargs = {'user_profile': {'read_only': True}}
 
 
-class QuestionSerializer(serializers.ModelSerializer):
-    """Serializes questions"""
-
-    class Meta:
-        model = models.Question
-        fields = ('id', 'created_on', 'topic', 'subtopic', 'dependencies',
-                  'question', 'correctAnswers', 'appendix', 'hint', 'imageSrc',
-                  'user_profile', 'validation')
-        extra_kwargs = {'user_profile': {'read_only': True}, 'appendix': {'required': False}}
-
-
 class AnswerSerializer(serializers.ModelSerializer):
     """Serializes answers"""
